Remove custom precondition; use Discord permissions
